# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sahayak AI Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/search/vector": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Vector Search",
        "operationId": "vector_search_search_vector_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VectorSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Vector Search Search Vector Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/search/rag": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Rag Search",
        "operationId": "rag_search_search_rag_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RagSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Rag Search Search Rag Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/voice/transcribe": {
      "post": {
        "summary": "Transcribe Voice",
        "description": "Transcribes an audio file using the existing Whisper model.",
        "operationId": "transcribe_voice_voice_transcribe_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_transcribe_voice_voice_transcribe_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscribeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/voice/speak": {
      "post": {
        "summary": "Speak Text",
        "description": "Converts text to speech using the gTTS library and returns MP3 audio bytes.",
        "operationId": "speak_text_voice_speak_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeakRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "audio/mpeg": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/voice/voice_query": {
      "post": {
        "summary": "Voice Query",
        "description": "Pipeline: transcribe audio -> RAG query -> speak answer.",
        "operationId": "voice_query_voice_voice_query_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Language"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_voice_query_voice_voice_query_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_transcribe_voice_voice_transcribe_post": {
        "properties": {
          "audio_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Audio File"
          }
        },
        "type": "object",
        "required": [
          "audio_file"
        ],
        "title": "Body_transcribe_voice_voice_transcribe_post"
      },
...
      "Body_voice_query_voice_voice_query_post": {
        "properties": {
          "audio_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Audio File"
          }
        },
        "type": "object",
        "required": [
          "audio_file"
        ],
        "title": "Body_voice_query_voice_voice_query_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RagSearchRequest": {
        "properties": {
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "description": "The search query text"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Alternative to query"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "description": "Number of results to return",
            "default": 5
          },
          "target": {
            "type": "string",
            "title": "Target",
            "description": "Target collection or 'auto'",
            "default": "auto"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Optional session ID for chat history"
          },
          "learning_mode": {
            "type": "string",
            "title": "Learning Mode",
            "description": "Learning mode context",
            "default": "student"
          },
          "user_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Mode",
            "description": "User-defined mode context"
          }
        },
        "type": "object",
        "title": "RagSearchRequest"
      },
...
      "SpeakRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SpeakRequest"
      },
...
      "TranscribeResponse": {
        "properties": {
          "transcribed_text": {
            "type": "string",
            "title": "Transcribed Text"
          }
        },
        "type": "object",
        "required": [
          "transcribed_text"
        ],
        "title": "TranscribeResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VectorSearchRequest": {
        "properties": {
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "description": "The search query text"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Alternative to query"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "description": "Number of results to return",
            "default": 5
          },
          "target": {
            "type": "string",
            "title": "Target",
            "description": "Target collection or 'auto'",
            "default": "auto"
          }
        },
        "type": "object",
        "title": "VectorSearchRequest"
      },
      "VoiceQueryResponse": {
        "properties": {
          "text_answer": {
            "type": "string",
            "title": "Text Answer"
          },
          "audio_answer": {
            "type": "string",
            "title": "Audio Answer"
          },
          "sources": {
            "items": {},
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "text_answer",
          "audio_answer",
          "sources"
        ],
        "title": "VoiceQueryResponse"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}